Optimize sine sampling with a lookup table
Before it took about 1/2 an iteration, now it's maybe 1/4.

+++ i2ctest/SineWave.hpp
@@ -6,7 +6,7 @@ class SineWave {
 public:
     SineWave(double freq);
     void setFrequency(double newfreq, double time);
-    double sample(double time);
+    uint16_t sample(double time);
 
 private:
     double freq;
